Further file restructuring
