+++ hpcrocket/ssh/sshexecutor.py
@@ -1,43 +1,12 @@
-import os
-from dataclasses import dataclass, replace
 from typing import List, Optional
 
 import paramiko as pm
 import paramiko.channel as channel
 from hpcrocket.core.executor import (CommandExecutor, CommandExecutorFactory,
                                      RunningCommand)
+
+from hpcrocket.ssh.connectiondata import ConnectionData
 from hpcrocket.ssh.errors import SSHError
-
-
-@dataclass
-class ConnectionData:
-    hostname: str
-    username: str
-    password: Optional[str] = None
-    keyfile: Optional[str] = None
-    key: Optional[str] = None
-    port: int = 22
-
-    @staticmethod
-    def with_resolved_keyfile(connection_data: 'ConnectionData') -> 'ConnectionData':
-        return ConnectionData._resolve_keyfile_in_connection(connection_data)
-
-    @staticmethod
-    def _resolve_keyfile_in_connection(connection):
-        keyfile = ConnectionData._resolve_keyfile_from_home_dir(connection.keyfile)
-        connection = replace(connection, keyfile=keyfile)
-        return connection
-
-    @staticmethod
-    def _resolve_keyfile_from_home_dir(keyfile: str) -> Optional[str]:
-        home_dir = os.environ["HOME"]
-        if not keyfile:
-            return None
-
-        if keyfile.startswith("~/"):
-            keyfile = keyfile.replace("~/", home_dir + "/", 1)
-
-        return keyfile
 
 
 class RemoteCommand(RunningCommand):
@@ -116,6 +85,7 @@ class SSHExecutorFactory(CommandExecutorFactory):
         executor = SSHExecutor()
         executor.connect(connection, proxyjumps=proxyjumps)
         return executor
+
 
 def build_channel_with_proxyjumps(connection: ConnectionData, proxyjumps: List[ConnectionData]) -> Optional[pm.Channel]:
     channel = None
